pyproject.toml
[project]
name = "mcp-server-for-cbt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.3",
    "langchain-groq>=1.1.0",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.4",
    "mcp-python>=0.1.4",
    "mcp[cli]>=1.23.3",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
